refactor(admin): consolide la fiche locataire sous /admin/people/tenants/[id]
Phase 3 du nettoyage admin : la fiche locataire vivait a /admin/tenants/[id]
alors que les fiches proprietaire/prestataire vivaient sous /admin/people/.
Cette incoherence imposait un getDetailUrl() avec un cas special et un
stub de redirection /admin/people/tenants/[id] -> /admin/tenants/[id].

- Deplace app/admin/tenants/[id]/* vers app/admin/people/tenants/[id]/
  (ecrase le stub de redirection) ; /admin/tenants/ est desormais vide
- PeopleClient.getDetailUrl() simplifie : toutes les fiches suivent
  /admin/people/{role}/{id}
- next.config.js : ajoute le redirect 301 /admin/tenants/:id ->
  /admin/people/tenants/:id pour ne pas casser les liens existants
- Header admin : retire le pattern /admin/tenants devenu inutile
